Add initial grpkg scripts
These are experimental scripts to assist in the development
of addons for Gramps.  The intent is for developers to be able
to use them to start, maintain, and submit their addons.  Note
that this primarily assists with PRs to the addons-source repository,
and in submitting PRs to the addons (packaged form) repository, but
does _not_ manage either upstream repository..

Signed-off-by: Al Stone <ahs3@ahs3.net>