Add a workspace creation commmand
The command "./make.py gramps60 workplace MyAddon somepath" will
create the directory 'somepath', git clone the gramps, addons-source,
and addons into 'somepath', create the needed 'MyAddOn' directory in
addons-source, and populate it with templates for the required files.

This does not handle localization yet.

A small config file is created and used to track git user info,
as well.

Signed-off-by: Al Stone <ahs3@ahs3.net>